refactor move element handler
